refactor(gui): split buildMenuBar into MainWindow_Menus.cpp (#3351 Phase 1b)
Third PR of the #3351 monolith-decomposition series. Pure code motion —
no behavior change, no MainWindow.h changes.

Moves buildMenuBar() (1,048 lines: every QMenu/QAction, their
enable/disable wiring, and ~70 inline connect lambdas) out of
MainWindow.cpp into the new MainWindow_Menus.cpp.

One supporting promotion, called out for review:

- The keyboard-shortcut mutable state (s_keyboardShortcutsEnabled,
  s_sliderShortcutLeaseActive) and its accessor helpers
  (textInputCaptured, shortcutInputCaptured, shortcutGuard,
  leaseHolderBusy) gain external linkage, declared in a new internal
  header MainWindowShortcutState.h. The View-menu shortcut toggle —
  which writes the flag and hands shortcutGuard to
  ShortcutManager::rebuildShortcuts — moved with buildMenuBar, so the
  state now has writers in two TUs. Definitions stay in MainWindow.cpp;
  the header documents why file-scope state exists at all (the
  ShortcutManager guard is a receiver-less std::function that cannot
  read MainWindow members). The Shortcuts TU (Phase 1c) will take
  ownership of the definitions.

Breadcrumb comment at the excision site points to the new file.

MainWindow.cpp: 17,308 → 16,261 lines (series total: 19,474 → 16,261,
−16.5%).
Verified: full build clean, all 32 ctest suites pass, a11y lint clean.

Refs #3351. Principle XI.
